Refactor Groq trip plan sanitization into dedicated sanitizer
